histogram popup enhancements - open histogram next to the viewer - show alpha channel value in label - shortcut key can open flipbook histogram - fix rect-picking 16bpc image overflow
 every popup commands now hide popup when it is already shown

+++ toonz/sources/tnztools/stylepicker.cpp
@@ -261,7 +261,7 @@ TPixel64 StylePicker::pickAverageColor16(const TRectD &rect) const {
   assert(raster64);
   if (!raster64) return TPixel64::Transparent;
 
-  UINT r = 0, g = 0, b = 0, m = 0, size = 0;
+  uint64_t r = 0, g = 0, b = 0, m = 0, size = 0;
   for (int y = topLeft.y; y < bottomRight.y; y++) {
     TPixel64 *p = &raster64->pixels(y)[topLeft.x];
     for (int x = topLeft.x; x < bottomRight.x; x++, p++) {
